requirements.txt
asgiref==3.2.10
django-crispy-forms==1.9.2
django-widget-tweaks==1.4.8
django==3.1
djangorestframework==3.11.1
pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
